Refactor to simplify exported files
